+++ src/core/request/interface.ts
@@ -13,6 +13,7 @@ import type Cache from 'core/cache/interface';
 import type Response from 'core/request/response';
 import type { ResponseType } from 'core/request/response';
 
+import type RequestError from 'core/request/error';
 import type RequestContext from 'core/request/context';
 
 import type { defaultRequestOpts } from 'core/request/const';
@@ -257,6 +258,11 @@ export interface CreateRequestOptions<D = unknown> {
 	readonly method?: RequestMethod;
 
 	/**
+	 * Options to retry bad requests
+	 */
+	retry?: RetryOptions | number;
+
+	/**
 	 * Mime type of request data (if not specified, it will be casted dynamically)
 	 */
 	contentType?: string;
@@ -407,4 +413,24 @@ export interface CreateRequestOptions<D = unknown> {
 	 * Custom request engine
 	 */
 	engine?: RequestEngine;
+}
+
+/**
+ * Options to retry bad requests
+ * @typeparam D - response data type
+ */
+export interface RetryOptions<D = unknown> {
+	/**
+	 * Maximum number of attempts to request
+	 */
+	attempts?: number;
+
+	/**
+	 * Returns a number in milliseconds (or a promise) to wait before the next attempt.
+	 * If the function returns false, it will prevent all further attempts.
+	 *
+	 * @param attempt - current attempt number
+	 * @param error - error object
+	 */
+	delay?(attempt: number, error: RequestError<D>): number | Promise<void> | false;
 }